server/requirements.txt
openai
websocket
'uvicorn[standard]'
python-dotenv